% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metodo_tukey.R
\name{calcular_A}
\alias{calcular_A}
\title{Cálculo del valor A en intervalos de confianza para comparación de medias}
\usage{
calcular_A(alfa, r, sigma)
}
\arguments{
\item{alfa}{Nivel de significancia deseado.}

\item{r}{Número de réplicas por tratamiento.}

\item{sigma}{Desviación estándar estimada de la variable de interés.}
}
\value{
Un valor numérico que representa \code{A}, la mitad de la longitud esperada
del intervalo de confianza.
}
\description{
Esta función calcula el valor \code{A}, que representa una estimación ajustada
de la mitad de la longitud esperada del intervalo de confianza para comparar dos medias
en un diseño experimental. Es útil para evaluar si el número de réplicas \code{r}
genera un intervalo suficientemente estrecho, es decir, si \code{2A ≈ D}.
}
\examples{
calcular_A(alfa = 0.05, r = 6, sigma = sqrt(141.6))

}
