% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metodo_tukey.R
\name{calcular_r_B}
\alias{calcular_r_B}
\title{Cálculo del número de réplicas según fórmula básica corregida}
\usage{
calcular_r_B(S1_sq, df1, df2, beta, alpha, t, d)
}
\arguments{
\item{S1_sq}{Estimación de la varianza experimental (\code{S1^2}).}

\item{df1}{Grados de libertad del estimador de varianza experimental.}

\item{df2}{Grados de libertad del error.}

\item{beta}{Nivel de error tipo II (es decir, \code{1 - potencia deseada}).}

\item{alpha}{Nivel de significancia para las comparaciones múltiples.}

\item{t}{Número de tratamientos.}

\item{d}{Diferencia mínima detectable entre medias.}
}
\value{
Un valor numérico que representa el número estimado de réplicas \code{r} (no redondeado).
}
\description{
Esta función estima el número requerido de réplicas \code{r} en un diseño experimental
con múltiples tratamientos, usando una versión corregida de la fórmula 5.24. Se basa en
cuantiles de la distribución F (para potencia estadística) y de la distribución de Tukey
(para comparaciones múltiples), así como en la varianza experimental estimada.
}
\examples{
# Ejemplo directo con valores definidos
calcular_r_B(
  S1_sq = 141.6,
  df1 = 40,
  df2 = 48,
  beta = 0.1,
  alpha = 0.1,
  t = 6,
  d = 20
)

# Ejemplo extendido (como prueba)
S1_sq_ <- 141.6
Df1 <- 40
D <- 20
Df2 <- 48
Beta <- 0.1
Alpha <- 0.1
T_ <- 6

qf(1 - 0.10, df1 = 48, df2 = 40)
qtukey(p = 0.90, nmeans = 5, df = 48)
ptukey(q = 4.2, nmeans = 5, df = 48)

calcular_r_B(
  S1_sq = S1_sq_,
  df1 = Df1,
  d = D,
  df2 = Df2,
  beta = Beta,
  alpha = Alpha,
  t = T_
)

}
