% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Costos.R
\name{numero_de_tratamientos_y_replicas_con_efectos_aleatorios}
\alias{numero_de_tratamientos_y_replicas_con_efectos_aleatorios}
\title{Cálculo del número de tratamientos y réplicas bajo un modelo de efectos aleatorios}
\usage{
numero_de_tratamientos_y_replicas_con_efectos_aleatorios(
  costo_tratamiento,
  costo_ue,
  sigma_cuadrado,
  rho,
  v_max
)
}
\arguments{
\item{costo_tratamiento}{Costo unitario de cada tratamiento (\eqn{C_1}).}

\item{costo_ue}{Costo unitario por cada unidad experimental (\eqn{C_2}).}

\item{sigma_cuadrado}{Varianza de los errores (\eqn{\sigma^2}).}

\item{rho}{Proporción de la varianza total atribuida a los efectos aleatorios (\eqn{\rho = \sigma_\tau^2 / \sigma^2}).}

\item{v_max}{Máximo valor permitido para la varianza relativa (\eqn{v_{\max}}).}
}
\value{
Una lista con dos elementos:
\describe{
\item{\code{num_de_tratamientos}}{Número estimado de tratamientos (entero redondeado).}
\item{\code{num_de_replicas}}{Número estimado de réplicas por tratamiento (entero redondeado).}
}
}
\description{
Esta función estima el número óptimo de tratamientos y el número de réplicas por tratamiento
en un diseño experimental con modelo de efectos aleatorios. Se consideran los costos de tratamientos y unidades experimentales,
la varianza de error, la proporción de varianza atribuible a los efectos aleatorios (\code{rho}),
y un valor máximo permitido para la varianza relativa de las medias (\code{v_max}).
}
\details{
El procedimiento utiliza las siguientes fórmulas:
\itemize{
\item Se estima la varianza de efectos aleatorios:
\deqn{\sigma^2_A = \rho \sigma^2}
\item Número de tratamientos:
\deqn{a = \left(\frac{1}{v_{\max}}\right) \left(\sigma^2_A + \sqrt{\frac{\sigma^2_A \sigma^2 C_2}{C_1}}\right)}
\item Número de réplicas por tratamiento:
\deqn{r = \sqrt{\frac{\sigma^2 C_1}{\sigma^2_A C_2}}}
}
}
\examples{
resultado <- numero_de_tratamientos_y_replicas_con_efectos_aleatorios(
  costo_tratamiento = 150000,
  costo_ue = 50000,
  sigma_cuadrado = 416.21,
  rho = 0.3796,
  v_max = 43.49
)
resultado

}
