% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Costos.R
\name{proporcionalidad_con_costo_ni_tamano_de_muestra}
\alias{proporcionalidad_con_costo_ni_tamano_de_muestra}
\title{Asignación proporcional de réplicas considerando costos y presupuesto total}
\usage{
proporcionalidad_con_costo_ni_tamano_de_muestra(a, sigmas, costos, costo_total)
}
\arguments{
\item{a}{Número de tratamientos.}

\item{sigmas}{Vector numérico de desviaciones estándar estimadas para cada tratamiento (longitud igual a \code{a}).}

\item{costos}{Vector numérico de costos unitarios por tratamiento (longitud igual a \code{a}).}

\item{costo_total}{Presupuesto total disponible para realizar las observaciones.}
}
\value{
Un vector numérico de longitud \code{a} indicando el número de réplicas ajustadas proporcionalmente para cada tratamiento,
redondeadas al número entero más cercano.
}
\description{
Esta función calcula un esquema de asignación proporcional de réplicas por tratamiento,
considerando tanto las desviaciones estándar como los costos unitarios de observación
de cada tratamiento. La asignación se ajusta para respetar un presupuesto total disponible (\code{costo_total}).
}
\details{
El procedimiento de asignación sigue los siguientes pasos:
\itemize{
\item Se define \eqn{\lambda = 1/a}.
\item Se calcula el factor \eqn{\phi}:
\deqn{\phi = \left( \sum_{i=1}^a \lambda \sigma_i \sqrt{c_i} \right)^2 / (costo\_total)^2}
\item Finalmente, el número de réplicas para cada tratamiento se estima como:
\deqn{r_i = \mathrm{round}\left(\frac{\lambda \sigma_i}{\sqrt{\phi c_i}}\right)}
}

Si la longitud de \code{sigmas} o \code{costos} no coincide con \code{a}, se detiene la ejecución con un mensaje de error.
}
\examples{
sigmas <- c(6.27, 9.57, 12, 3.32)
costos <- c(1000, 200, 700, 1100)
costo_total <- 50000
proporcionalidad_con_costo_ni_tamano_de_muestra(a = 4, sigmas = sigmas, costos = costos, costo_total = costo_total)

}
