% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metodo_tukey.R
\name{revision_intervalo}
\alias{revision_intervalo}
\title{Revisión de intervalo para número óptimo de réplicas}
\usage{
revision_intervalo(info_r, des = 5)
}
\arguments{
\item{info_r}{Lista resultante de la función \code{calcular_df2}, que debe contener
los elementos \code{r}, \code{dif}, \code{alfa}, y \code{S1}.}

\item{des}{Desviación entera para explorar alrededor del valor de \code{r}. Por defecto es 5.}
}
\value{
Una lista con los siguientes elementos:
\describe{
\item{r_}{Valor de \code{r} que minimiza la diferencia entre \code{2A} y \code{d}.}
\item{A}{Valor de \code{A} correspondiente a \code{r_}.}
\item{r_l}{Vector de valores de \code{r} evaluados.}
\item{A_l}{Vector de valores de \code{A} correspondientes a cada \code{r_l}.}
\item{position}{Índice de \code{r_} dentro del vector evaluado.}
}
}
\description{
Esta función evalúa un rango de valores alrededor del número estimado de
réplicas \code{r}, calculando el valor correspondiente de \code{A} para cada uno
y determinando cuál valor de \code{r} hace que la longitud del intervalo
de confianza (\code{2A}) sea más cercana a la diferencia mínima detectable \code{d}.
}
\examples{

# Primero es necesario ejecutar calcular_r_MT
info <- calcular_r_MT(
  T_ = 6, D = 20, ro = 5,
  S1 = sqrt(141.6), df1 = 40,
  alfa = 0.05, Beta = 0.1
)
revision_intervalo(info_r = info, des = 3)

}
